+++ NEWS.md
@@ -5,4 +5,8 @@ New features
 ------------
 
 * `gh_browse()` opens up a GitHub repository in a web browser (#4)
+* `gh_label_clone()` copies labels from another repo (#2)
+* `gh_label_create()` creates a label (#2)
+* `gh_label_delete()` deletes label(s) (#2)
+* `gh_label_list()` lists label(s) (#2)
 * `gh_repo_edit()` edits repository settings (#6)